% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_quiz.R
\name{generate_quiz}
\alias{generate_quiz}
\title{Regenerate Quiz Content}
\usage{
generate_quiz(filepath = "quiz/tidy.Rmd")
}
\arguments{
\item{filepath}{Character string of file path to R Markdown file}
}
\description{
Re-render and re-run the R Markdown file for the tidyverse quiz
}
\examples{
\dontrun{
generate_quiz()
}
}
